Use size_t instead of int.

+++ mkpass.cc
@@ -7,7 +7,7 @@
 // https://stackoverflow.com/questions/47977829/generate-a-random-string-in-c11
 // g++ -std=c++17 -Wall -Wextra -pedantic -o mkpass mkpass.cc
 
-std::string random_string(int size) {
+std::string random_string(size_t size) {
     std::string str(
         "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz");
